move log decorator to its own file (#157)
