+++ window.go
@@ -1,9 +1,5 @@
 package app
 
-import (
-	"fmt"
-)
-
 // Window is the interface that describes a window.
 type Window interface {
 	Navigator
@@ -141,181 +137,3 @@ const (
 	VibeMediumLight
 	VibeUltraDark
 )
-
-type windowWithLogs struct {
-	Window
-}
-
-func (w *windowWithLogs) WhenWindow(f func(Window)) {
-	f(w)
-}
-
-func (w *windowWithLogs) WhenNavigator(f func(Navigator)) {
-	f(w)
-}
-
-func (w *windowWithLogs) Load(url string, v ...interface{}) {
-	parsedURL := fmt.Sprintf(url, v...)
-
-	WhenDebug(func() {
-		Debug("window %s is loading %s",
-			w.ID(),
-			parsedURL,
-		)
-	})
-
-	w.Window.Load(url, v...)
-	if w.Err() != nil {
-		Log("window %s failed to load %s: %s",
-			w.ID(),
-			parsedURL,
-			w.Err(),
-		)
-	}
-}
-
-func (w *windowWithLogs) Render(c Compo) {
-	WhenDebug(func() {
-		Debug("window %s is rendering %T",
-			w.ID(),
-			c,
-		)
-	})
-
-	w.Window.Render(c)
-	if w.Err() != nil {
-		Log("window %s failed to render %T: %s",
-			w.ID(),
-			c,
-			w.Err(),
-		)
-	}
-}
-
-func (w *windowWithLogs) Reload() {
-	WhenDebug(func() {
-		Debug("window %s is reloading", w.ID())
-	})
-
-	w.Window.Reload()
-	if w.Err() != nil {
-		Log("window %s failed to reload: %s",
-			w.ID(),
-			w.Err(),
-		)
-	}
-}
-
-func (w *windowWithLogs) Previous() {
-	WhenDebug(func() {
-		Debug("window %s is loading previous", w.ID())
-	})
-
-	w.Window.Previous()
-	if w.Err() != nil {
-		Log("window %s failed to load previous: %s",
-			w.ID(),
-			w.Err(),
-		)
-	}
-}
-
-func (w *windowWithLogs) Next() {
-	WhenDebug(func() {
-		Debug("window %s is loading next", w.ID())
-	})
-
-	w.Window.Next()
-	if w.Err() != nil {
-		Log("window %s failed to load next: %s",
-			w.ID(),
-			w.Err(),
-		)
-	}
-}
-
-func (w *windowWithLogs) Close() {
-	WhenDebug(func() {
-		Debug("window %s is closing", w.ID())
-	})
-
-	w.Window.Close()
-	if w.Err() != nil {
-		Log("window %s failed to close: %s",
-			w.ID(),
-			w.Err(),
-		)
-	}
-}
-
-func (w *windowWithLogs) Move(x, y float64) {
-	WhenDebug(func() {
-		Debug("window %s is moving to x:%.2f y:%.2f",
-			w.ID(),
-			x,
-			y,
-		)
-	})
-
-	w.Window.Move(x, y)
-}
-
-func (w *windowWithLogs) Center() {
-	WhenDebug(func() {
-		Debug("window %s is moving to center", w.ID())
-	})
-
-	w.Window.Center()
-}
-
-func (w *windowWithLogs) Resize(width, height float64) {
-	WhenDebug(func() {
-		Debug("window %s is resizing to width:%.2f height:%.2f",
-			w.ID(),
-			width,
-			height,
-		)
-	})
-
-	w.Window.Resize(width, height)
-}
-
-func (w *windowWithLogs) Focus() {
-	WhenDebug(func() {
-		Debug("window %s is getting focus", w.ID())
-	})
-
-	w.Window.Focus()
-}
-
-func (w *windowWithLogs) FullScreen() {
-	WhenDebug(func() {
-		Debug("window %s is entering full screen", w.ID())
-	})
-
-	w.Window.FullScreen()
-}
-
-func (w *windowWithLogs) ExitFullScreen() {
-	WhenDebug(func() {
-		Debug("window %s is exiting full screen", w.ID())
-	})
-
-	w.Window.ExitFullScreen()
-}
-
-func (w *windowWithLogs) Minimize() {
-	WhenDebug(func() {
-		Debug("window %s is minimizing", w.ID())
-	})
-
-	w.Window.Minimize()
-}
-
-func (w *windowWithLogs) Deminimze() {
-	WhenDebug(func() {
-		Debug("window %s is deminimizing", w.ID())
-	})
-
-	w.Window.Deminimize()
-}
